Show files content on console
First check if directory exists, creating it if not. Then clone or
pull repository to get latest files. After that, show files content
on console using git commands "ls-files", which list all files, and
"show HEAD:<path-to-file>", which prints latest version of file
regardless of which commit changed it last time.