Consume a real data source.
- Made DataSource read a real git repository, with
capability to return a log history. Git root can be passed
through constructor.

- Made DataSource and Repository implement AutoCloseable, so
it can close the IO with the git repo access when not in use
anymore. Calling repo.close() closes the DataSource which in
turn closes the Git repository.

- Added callback to main Composable so it closes the Repository
when the user closes the app.

- Generified the Composable that collect repositories flows,
and the Composable that takes an item List (usually from
the composable that collects the flow).